fix swapped longan_led_on/off - board has led fixed at 3.3v
if we check out the schematic[0] we see that the LEDs are hardwired to 3.3V,
we cannot control the anode only the negative cathode. So writing out a LOW
value to the GPIO port connected with a LED (PC13, PA1, PA2) will make it light
up, and writing out a HIGH value will interrupt current flow and turn the LED off.

In the example this seemed to work as the loop did both, switch on and switch
off, so the swapped logic was seemingly no issue. But if one just test the
following main before my change, they can see that it was broken:

int main(void) {
    longan_led_init();
    longan_led_off(); // LED should be out, but it will light!

    while (1);
}

As people will use this example to start developing their own things,
it really should be correct!

[0]: http://dl.sipeed.com/LONGAN/Nano/HDK/Longan%20Nano%202663/Longan%20nano%202663(Schematic).pdf